# backend-python/openapi.yaml
openapi: 3.0.3
info:
  title: ResQBite API
  version: 1.0.0
  description: Food rescue and delivery backend for NGOs, donors and volunteers.
paths:
...
  /api/donations{id}/recommendations/:
    get:
      operationId: donations_recommendations_retrieve_2
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - donations
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Donation'
          description: ''
...
  /api/donations/{id}/recommendations/:
    get:
      operationId: donations_recommendations_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - donations
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Donation'
          description: ''
...
components:
  schemas:
    CategoryEnum:
      enum:
      - VEG
      - NON_VEG
      - VEGAN
      - DESSERTS
      - BAKERY
      - BEVERAGES
      - FRUITS
      - PACKED_FOOD
      type: string
      description: |-
        * `VEG` - Veg
        * `NON_VEG` - Non-Veg
        * `VEGAN` - Vegan
        * `DESSERTS` - Desserts
        * `BAKERY` - Bakery
        * `BEVERAGES` - Beverages
        * `FRUITS` - Fruits
        * `PACKED_FOOD` - Packed Food
    Donation:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        donor:
          type: string
          format: uuid
          readOnly: true
        donor_name:
          type: string
          readOnly: true
        organization:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        food_name:
          type: string
          maxLength: 200
        category:
          $ref: '#/components/schemas/CategoryEnum'
        food_type:
          type: string
          maxLength: 100
        quantity:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        quantity_unit:
          type: string
          maxLength: 30
        people_served:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        prepared_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        pickup_address:
          type: string
          maxLength: 255
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        special_instructions:
          type: string
        status:
          allOf:
          - $ref: '#/components/schemas/DonationStatusEnum'
          readOnly: true
        food_safety:
          $ref: '#/components/schemas/FoodSafetyChecklist'
        images:
          type: array
          items:
            type: string
            format: uri
          writeOnly: true
        image_details:
          type: array
          items:
            $ref: '#/components/schemas/DonationImage'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - category
      - created_at
      - donor
      - donor_name
      - expires_at
      - food_name
      - food_safety
      - food_type
      - id
      - image_details
      - organization
      - prepared_at
      - quantity
      - status
      - updated_at
    DonationImage:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        url:
          type: string
          readOnly: true
        alt_text:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - alt_text
      - created_at
      - id
      - url
...
    DonationStatusEnum:
      enum:
      - PENDING
      - ACCEPTED
      - COMPLETED
      - CANCELLED
      type: string
      description: |-
        * `PENDING` - Pending
        * `ACCEPTED` - Accepted
        * `COMPLETED` - Completed
        * `CANCELLED` - Cancelled
    FoodSafetyChecklist:
      type: object
      properties:
        freshly_prepared:
          type: boolean
        properly_packed:
          type: boolean
        expiry_marked:
          type: boolean
        food_type_marked:
          type: boolean
        allergens:
          type: string
          maxLength: 500
      required:
      - expiry_marked
      - food_type_marked
      - freshly_prepared
      - properly_packed
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT